# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GymPal Backend API",
    "description": "Complete API for the GymPal fitness application with social features, personal tracking, workouts, and analytics",
    "version": "1.0.0",
    "contact": {
      "name": "GymPal Team",
      "email": "support@gympal.app",
      "url": "https://gympal.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.gympal.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/dashboard": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get dashboard data",
        "description": "Get dashboard data for the current user",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Dashboard data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/dashboard/stats": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get user dashboard statistics",
        "description": "Get user dashboard statistics",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "timeframe",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["week", "month", "year", "all"],
              "default": "month"
            }
          },
          {
            "name": "include_social",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/dashboard/recent-activity": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get recent user activity",
        "description": "Get recent user activity",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent activity retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/dashboard/workout-progress": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get workout progress",
        "description": "Get workout progress for the current user",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["week", "month", "year", "all"],
              "default": "month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workout progress retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/dashboard/analytics": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get user analytics",
        "description": "Get user analytics",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["week", "month", "year", "all"],
              "default": "month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/dashboard/leaderboard": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get user leaderboard",
        "description": "Get user leaderboard",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["workouts", "achievements", "goals", "posts", "likes", "comments", "followers", "following"],
              "default": "workouts"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/dashboard/calendar": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get user calendar data",
        "description": "Get user calendar data",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Calendar data retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "DashboardResponse": {
        "type": "object",
        "properties": {
          "stats": {
            "$ref": "#/components/schemas/DashboardStatsResponse"
          },
          "recentActivity": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "workoutProgress": {
            "type": "object"
          },
          "achievements": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "DashboardStatsResponse": {
        "type": "object",
        "properties": {
          "totalWorkouts": {
            "type": "integer",
            "example": 45
          },
          "totalExercises": {
            "type": "integer",
            "example": 120
          },
          "totalDuration": {
            "type": "integer",
            "example": 3600
          },
          "averageWorkoutDuration": {
            "type": "number",
            "example": 45.5
          },
          "streak": {
            "type": "integer",
            "example": 7
          },
          "socialActivity": {
            "type": "object",
            "properties": {
              "posts": {
                "type": "integer",
                "example": 15
              },
              "likes": {
                "type": "integer",
                "example": 50
              },
              "comments": {
                "type": "integer",
                "example": 25
              }
            }
          }
        }
      },
...
    }
  }
}